main completed, the save and load function does not work correctly

+++ main.java
@@ -10,7 +10,7 @@ public class main {
     private static int selection;
     private static SistemaControllo sw;
     private static MessageDigest message;
-    private static String sha256_original = "14b382abb8736639dce380ac2c808db5fe72b3250eb198199e8ec4ca42cfc96d";
+    protected static String sha256_original = "14b382abb8736639dce380ac2c808db5fe72b3250eb198199e8ec4ca42cfc96d";
 
     public static void main(String[] args){
         String[] let = new String[10];
@@ -34,6 +34,7 @@ public class main {
                             }else{
                                 System.out.println("Password are not equals .. Aborting operation");
                             }
+                            resetBuffer(let);
                             break;
                     case 2:
                             System.out.println("Showing archive: ");
@@ -42,33 +43,49 @@ public class main {
                     case 3:
                             System.out.print(sw.toString()+"\n\nSelect the cell: ");
                             selection = Integer.parseInt(bufferReader.readLine());
-                            if(!sw.isFree(selection)){
-                                sw.deleteNode(selection);
+                            if(selection > sw.getSize()){
+                                System.out.println("There i no cell in the selected position .. Aborting");
+                                break;
+                            }else{
+                                if(!sw.isFree(selection)){
+                                    sw.deleteNode(selection);
+                                    System.out.println("Cell successfuly deleted");
+                                }
                             }
                             break;
 
                     case 4:
                             System.out.print(sw.toString()+"\n\nSelect the cell: ");
                             selection = Integer.parseInt(bufferReader.readLine());
+                            if(selection > sw.getSize()){
+                                System.out.println("there is no existing cell in this position .. Aborting");
+                                break;
+                            }
                             System.out.print("\nYout want to get the password for "+ sw.getCell(selection)+" ?  (Y/N): ");
                             let[3] = bufferReader.readLine();
-                            if(let[3].equals("Y")){
+                            if(let[3].equalsIgnoreCase("Y")){
                                 System.out.print("Insert the password to unlock your page: ");
                                 let[3] = bufferReader.readLine();
                                 message = MessageDigest.getInstance("SHA-256");
-                                byte[] sha256 = message.digest(let[3].getBytes(StandardCharsets.UTF_8)); 
-                                if(sha256.equals(sha256_original.getBytes(StandardCharsets.UTF_8))){
+                                byte[] sha256 = message.digest(let[3].getBytes(StandardCharsets.UTF_8));
+                                let[9] = sw.byteToEx(sha256);
+                                if(let[9].equalsIgnoreCase(sha256_original)){
                                     System.out.println("Decrypted password: "+sw.decryptNode(selection));
                                 }
-                                message = null;
-                                sha256 = null;
-                                let[3] = null;
-                            }else{
-                                System.out.println("Aborting .. ");
+                                
                             }
+                            message = null;
+                            resetBuffer(let);
                             break;
                     case 5:
+
                             System.out.println("Saving the archive on the same origin path ");
+                            System.out.print("Path to save the File: ");
+                            let[0] = bufferReader.readLine();
+                            System.out.print("Path to save the Key: ");
+                            let[1] = bufferReader.readLine();
+                            System.out.print("Path to save the partial method: ");
+                            let[2] = bufferReader.readLine();
                             sw.serializeInfo(let[0], let[1],let[2]);
                             System.out.println(sw.toString()+"\n\n Archive successfuly saved! ..");
                             break;
@@ -80,6 +97,7 @@ public class main {
                             System.out.print("Path to load the partial method: ");
                             let[2] = bufferReader.readLine();
                             sw.loadData(let[0], let[1],let[2]);
+                            System.out.println(sw.toString()+"\n\n Archive successfuly loaded! ..");
                             break;
                     case 0:
                             start = 1;
@@ -93,7 +111,7 @@ public class main {
         }
     }
     public static void menu(){
-        System.out.println("Encrypt/Decrypt password archive\n");
+        System.out.println("\n\n--------------------------------\nEncrypt/Decrypt password archive\n");
         System.out.println("1 - new page");
         System.out.println("2 - show current archive");
         System.out.println("3 - delete page");
@@ -102,5 +120,10 @@ public class main {
         System.out.println("6 - Load");
         System.out.println("\n0 - Quit");
         System.out.print("\n\nselection: ");
+    }
+    public static void resetBuffer(String[] buffer){
+        for(int i=0;i<buffer.length;i++){
+            buffer[i] = null;
+        }
     }
 }
